refactor: authentication 폴더 분리
